+++ src/components/Home/BalanceBottomSheet.tsx
@@ -91,7 +91,7 @@ const BalanceBottomSheet = ({
       </View>
       <AppText
         style={tw`text-center text-xl font-bold tracking-tight text-slate-900 dark:text-gray-200 mt-4`}>
-        {t('home.profile.tickets.label')}
+        {t('home.profile.tickets.title')}
       </AppText>
       <AppText style={tw`text-left text-base font-normal text-slate-500 w-full mt-4`}>
         {t('home.profile.tickets.description')}